+++ pkgs/shelf/lib/shelf_io.dart
@@ -217,8 +217,9 @@ Future<void> _writeResponse(Response response, HttpResponse httpResponse) {
     httpResponse.headers.set(HttpHeaders.transferEncodingHeader, 'chunked');
   }
 
-  if (!response.headers.containsKey(HttpHeaders.serverHeader)) {
-    httpResponse.headers.set(HttpHeaders.serverHeader, 'dart:io with Shelf');
+  if (!response.headers.containsKey(_xPoweredByResponseHeader)) {
+    httpResponse.headers
+        .set(_xPoweredByResponseHeader, 'Dart with package:shelf');
   }
 
   if (!response.headers.containsKey(HttpHeaders.dateHeader)) {
@@ -229,6 +230,11 @@ Future<void> _writeResponse(Response response, HttpResponse httpResponse) {
       .addStream(response.read())
       .then((_) => httpResponse.close());
 }
+
+/// Common header to advertise the server technology being used.
+///
+/// See https://webtechsurvey.com/response-header/x-powered-by
+const _xPoweredByResponseHeader = 'X-Powered-By';
 
 // TODO(kevmoo) A developer mode is needed to include error info in response
 // TODO(kevmoo) Make error output plugable. stderr, logging, etc